fix: PR #417 R1 리뷰 반영 — lib layer, JSDoc, datePublished, 매직 넘버
- B1 lib/og.ts loadKoreanFont fetch 사이드 이펙트 → infrastructure/og/loadKoreanFont.ts로 이동
- S1/S2 og.ts·SymbolPageClient.tsx 멀티라인 JSDoc → 1줄 인라인 코멘트로 축약
- S3 news Article.datePublished new Date() → SITE_BUILD_DATE (lib/seo로 끌어올림). dateModified만 요청 시각 유지
- S4 sitemap 시간 매직 넘버 → MS_PER_HOUR / MS_PER_DAY (@/domain/constants/time). parseBuildDate 중복 제거 (lib/seo.SITE_BUILD_DATE 재사용)

거절: inline #1 (per-axis dynamic 의도), inline #2 (siglens-core 책임 영역 아님), S5 (글로벌 Footer landmark 중복 회피). 인라인 답변 + PR 코멘트로 사유 게시.

Validation: lint·tsc clean, 1274/1274 tests pass, build pass.
